% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addRegistryPackages}
\alias{addRegistryPackages}
\title{Add packages to registry.}
\usage{
addRegistryPackages(reg, packages)
}
\arguments{
\item{reg}{[\code{Registry}]\cr
Registry.}

\item{packages}{[\code{character}]\cr
Packages to add to registry.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{packages} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{batchExport}};
  \code{\link{batchUnexport}}; \code{\link{loadExports}};
  \code{\link{removeRegistryPackages}}
}

