% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRegistryFiles.R
\name{addRegistrySourceDirs}
\alias{addRegistrySourceDirs}
\title{Add source dirs to registry.}
\usage{
addRegistrySourceDirs(reg, src.dirs, src.now = TRUE)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{src.dirs}{[\code{character}]\cr
Paths to add to registry.
See \code{\link{makeRegistry}}.}

\item{src.now}{[\code{logical(1)}]
Source files now on master?
Default is \code{TRUE}.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{src.dirs} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}},
  \code{\link{addRegistrySourceFiles}},
  \code{\link{batchExport}}, \code{\link{batchUnexport}},
  \code{\link{loadExports}},
  \code{\link{removeRegistryPackages}},
  \code{\link{removeRegistrySourceDirs}},
  \code{\link{removeRegistrySourceFiles}},
  \code{\link{setRegistryPackages}}
}

