% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addRegistrySourceFiles}
\alias{addRegistrySourceFiles}
\title{Add source files to registry.}
\usage{
addRegistrySourceFiles(reg, src.files)
}
\arguments{
\item{reg}{[\code{Registry}]\cr
Registry.}

\item{src.files}{[\code{character}]\cr
Paths to add to registry.
See \code{\link{makeRegistry}}.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{src.files} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}};
  \code{\link{addRegistrySourceDirs}};
  \code{\link{batchExport}}; \code{\link{batchUnexport}};
  \code{\link{loadExports}};
  \code{\link{removeRegistryPackages}};
  \code{\link{removeRegistrySourceDirs}};
  \code{\link{removeRegistrySourceFiles}}
}

