\name{applyJobFunction}
\alias{applyJobFunction}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
applyJobFunction(reg, job, cache)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{job}{[\code{\link{Job}}]\cr Job.}

  \item{cache}{[\code{FileCache}]\cr Instance of
  \code{\link[BBmisc]{FileCache}}.}
}
\value{
[any]. Result of job.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

