% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{BatchJobs}
\alias{BatchJobs}
\alias{BatchJobs-package}
\title{The BatchJobs package}
\description{
The BatchJobs package
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/tudo-r/BatchJobs}}
  \item{Wiki:}{\url{https://github.com/tudo-r/BatchJobs/wiki}}
  \item{FAQ:}{\url{https://github.com/tudo-r/BatchJobs/wiki/FAQ}}
  \item{Configuration:}{\url{https://github.com/tudo-r/BatchJobs/wiki/Configuration}}
}

The package currently support the following further R options, which you can set
either in your R profile file or a script via \code{\link{options}}:

\describe{
  \item{BatchJobs.verbose}{This boolean flag can be set to \code{FALSE} to reduce the
    console output of the package operations. Usually you want to see this output in interactive
    work, but when you  use the package in e.g. knitr documents,
    it clutters the resulting document too much.}
  \item{BatchJobs.check.posix}{In its default setting, the package checks you registry file dir
  (and related user-defined directories) quite strictly to be POSIX compliant. Usually this a good
  idea, you do not want to have strange chars in your file paths, as this might results in problems
  when these paths get passed to the scheduler or other command-line tools that the package
  interoperates with. But on some OS this check might be too strict and cause problems.
  This boolean flag allows to disable the check entirely.}
}
}

