% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsHelpers.R
\name{cfReadBrewTemplate}
\alias{cfReadBrewTemplate}
\title{Cluster functions helper: Read in your brew template file.}
\usage{
cfReadBrewTemplate(template.file)
}
\arguments{
\item{template.file}{[\code{character(1)}]\cr
File path.}
}
\value{
[\code{character}].
}
\description{
This function is only intended for use in your own cluster functions implementation.

Simply reads your template and returns it as a character vector.
If you do this in the constructor of your cluster functions once, you can avoid this
repeated file access later on.
}

