% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIds.R
\name{checkIds}
\alias{checkIds}
\title{Check job ids.}
\usage{
checkIds(reg, ids, check.present = TRUE, len = NULL)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Vector of job ids.}

\item{check.present}{[\code{logical(1)}]\cr
Check if the ids are present in the database? Default is \code{TRUE}.}

\item{len}{[\code{integer(1)}]\cr
Expected length of id vector. Passed to \code{\link[checkmate]{asInteger}}.}
}
\value{
Invisibly returns the vector of ids, converted to integer.
}
\description{
Simply checks if probided vector of job ids is valid and
throws an error if something is odd.
}

