% Generated by roxygen2 (4.0.1): do not edit by hand
\name{configuration}
\alias{.BatchJobs.R}
\alias{configuration}
\title{BatchJobs configuration.}
\description{
In order to understand how the package should be configured
please read
\url{https://github.com/tudo-r/BatchJobs/wiki/Configuration}.
}

