% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testJob.R
\name{copyRequiredJobFiles}
\alias{copyRequiredJobFiles}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
copyRequiredJobFiles(reg1, reg2, id)
}
\arguments{
\item{reg1}{[\code{\link{Registry}}]\cr
Source registry.}

\item{reg2}{[\code{\link{Registry}}]\cr
Detination registry.}

\item{id}{[\code{character(1)}]\cr
Job id.}
}
\value{
Nothing.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

