% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbCreateJobDefTable}
\alias{dbCreateJobDefTable}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
dbCreateJobDefTable(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}
}
\value{
Nothing.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

