\name{dbGetJobs}
\alias{dbGetJobs}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
dbGetJobs(reg, ids)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{ids}{[\code{integer}]\cr Ids of selected jobs.}
}
\value{
[list of \code{\link{Job}}]. Retrieved jobs from DB.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

