\name{findDone}
\alias{findDone}
\title{Find jobs for which results are available.}
\usage{
  findDone(reg, ids)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{ids}{[\code{integer}]\cr Subset of job ids to
  restrict the result to.  Default is all jobs.}
}
\value{
  [\code{integer}]. Ids of jobs.
}
\description{
  Find jobs for which results are available.
}

