\name{findErrors}
\alias{findErrors}
\title{Find jobs where errors occured.}
\usage{
  findErrors(reg, ids)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{ids}{[\code{integer}]\cr Subset of job ids to
  restrict the result to.  Default is all jobs.}
}
\value{
  [\code{integer}]. Ids of jobs.
}
\description{
  Find jobs where errors occured.
}

