% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStatus.R
\name{findDone}
\alias{findDone}
\alias{findNotDone}
\alias{findMissingResults}
\alias{findErrors}
\alias{findNotErrors}
\alias{findTerminated}
\alias{findNotTerminated}
\alias{findSubmitted}
\alias{findNotSubmitted}
\alias{findOnSystem}
\alias{findNotOnSystem}
\alias{findRunning}
\alias{findNotRunning}
\alias{findStarted}
\alias{findNotStarted}
\alias{findExpired}
\alias{findDisappeared}
\title{Find jobs depending on computional state.}
\usage{
findDone(reg, ids, limit = NULL)

findNotDone(reg, ids, limit = NULL)

findMissingResults(reg, ids, limit = NULL)

findErrors(reg, ids, limit = NULL)

findNotErrors(reg, ids, limit = NULL)

findTerminated(reg, ids, limit = NULL)

findNotTerminated(reg, ids, limit = NULL)

findSubmitted(reg, ids, limit = NULL)

findNotSubmitted(reg, ids, limit = NULL)

findOnSystem(reg, ids, limit = NULL)

findNotOnSystem(reg, ids, limit = NULL)

findRunning(reg, ids, limit = NULL)

findNotRunning(reg, ids, limit = NULL)

findStarted(reg, ids, limit = NULL)

findNotStarted(reg, ids, limit = NULL)

findExpired(reg, ids, limit = NULL)

findDisappeared(reg, ids, limit = NULL)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Subset of job ids to restrict the result to.
Default is all jobs.}

\item{limit}{[\code{integer(1)}]\cr
Limit the number of returned ids.
Default is all ids.}
}
\value{
[\code{integer}]. Ids of jobs.
}
\description{
\code{findDone}: Find jobs which succesfully terminated.
}
