\name{getConfig}
\alias{getConfig}
\title{Returns a list of BatchJobs configuration settings}
\usage{
getConfig()
}
\value{
\code{list} of current configuration variables with classs
\dQuote{Config}.
}
\description{
Returns a list of BatchJobs configuration settings
}
\seealso{
\code{\link{loadConfig}}, \code{\link{setConfig}}
}

