% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getErrorMessages}
\alias{getErrorMessages}
\title{Get error messages of jobs.}
\usage{
getErrorMessages(reg, ids)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all jobs with errors.}
}
\value{
[\code{character}]. Error messages for jobs as character vector\cr
  \code{NA} if job has terminated successfully.
}
\description{
Get error messages of jobs.
}
\seealso{
\code{\link{showLog}} for more contextual error informations.
}

