% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\name{getJobIds}
\alias{getJobIds}
\title{Get ids of jobs in registry.}
\usage{
getJobIds(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}
}
\value{
[\code{character}].
}
\description{
Get ids of jobs in registry.
}

