% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJobLocation.R
\name{getJobLocation}
\alias{getJobLocation}
\title{Get the physical location of job files on the hard disk.}
\usage{
getJobLocation(reg, ids)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all jobs.}
}
\value{
[\code{character}] Vector of directories.
}
\description{
Get the physical location of job files on the hard disk.
}

