% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\name{getJobNr}
\alias{getJobNr}
\title{Get number of jobs in registry.}
\usage{
getJobNr(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}
}
\value{
[\code{integer(1)}].
}
\description{
Get number of jobs in registry.
}

