% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resources.R
\name{getJobResources}
\alias{getJobResources}
\title{Function to get the resources that were submitted for some jobs.}
\usage{
getJobResources(reg, ids, as.list = TRUE)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all submitted jobs.}

\item{as.list}{[\code{integer(1)}]\cr
If \code{FALSE} a data.frame will be returned.
Default is \code{TRUE}.}
}
\value{
[\code{list} | \code{data.frame}]. List (or data.frame) of resource lists as passed to \code{\link{submitJobs}}.
}
\description{
Throws an error if call it for unsubmitted jobs.
}

