\name{loadConfig}
\alias{loadConfig}
\title{Load a specific configuration file.}
\usage{
  loadConfig(conffile = ".BatchJobs.R")
}
\arguments{
  \item{conffile}{[\code{character(1)}]\cr Location of the
  configuration file to load.}
}
\value{
  Nothing.
}
\description{
  Load a specific configuration file.
}

