% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Registry.R
\name{loadRegistry}
\alias{loadRegistry}
\title{Load a previously saved registry.}
\usage{
loadRegistry(file.dir, work.dir)
}
\arguments{
\item{file.dir}{[\code{character(1)}]\cr
Location of the file.dir to load the registry from.}

\item{work.dir}{[\code{character(1)}]\cr
Location of the work. Unchanged if missing.}
}
\value{
[\code{\link{Registry}}].
}
\description{
Load a previously saved registry.
}

