% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsSLURM.R
\name{makeClusterFunctionsSLURM}
\alias{makeClusterFunctionsSLURM}
\title{Create cluster functions for SLURM-based systems.}
\usage{
makeClusterFunctionsSLURM(template.file, list.jobs.cmd = c("squeue", "-h",
  "-o \%i", "-u $USER"), list.job.line.skip = 0L, cluster.name)
}
\arguments{
\item{template.file}{[\code{character(1)}]\cr
Path to a brew template file that is used for the PBS job file.}

\item{list.jobs.cmd}{[\code{character}]\cr
Change default system command / options to list jobs.
The first entry is the command, the following the options.
See \code{\link[BBmisc]{system3}}.}

\item{list.job.line.skip}{[\code{integer(1)}]\cr
Change how many lines of the job list should be skipped. Can be useful if \code{squeue} is giving
additional output.}

\item{cluster.name}{[\code{character(1)}]\cr
If an additional cluster name has to be specified for listing or deleting jobs it can be 
supplied with this argument. it will be added as \code{--clusters=cluster.name} for SLURM.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Job files are created based on the brew template
\code{template.file}. This file is processed with brew and then
submitted to the queue using the \code{sbatch} command. Jobs are
killed using the \code{scancel} command and the list of running jobs
is retrieved using \code{squeue}. The user must have the
appropriate privileges to submit, delete and list jobs on the
cluster (this is usually the case).

The template file can access all arguments passed to the
\code{submitJob} function, see here \code{\link{ClusterFunctions}}.
It is the template file's job to choose a queue for the job
and handle the desired resource allocations.
Examples can be found on
\url{https://github.com/tudo-r/BatchJobs/tree/master/examples/cfSLURM}.
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsLocal}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}}
}

