% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRegistryFiles.R
\name{removeRegistrySourceDirs}
\alias{removeRegistrySourceDirs}
\title{Remove packages from registry.}
\usage{
removeRegistrySourceDirs(reg, src.dirs)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{src.dirs}{[\code{character}]\cr
Paths to remove from registry.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{src.dirs} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}},
  \code{\link{addRegistrySourceDirs}},
  \code{\link{addRegistrySourceFiles}},
  \code{\link{batchExport}}, \code{\link{batchUnexport}},
  \code{\link{loadExports}},
  \code{\link{removeRegistryPackages}},
  \code{\link{removeRegistrySourceFiles}},
  \code{\link{setRegistryPackages}}
}

