% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRegistryFiles.R
\name{removeRegistrySourceFiles}
\alias{removeRegistrySourceFiles}
\title{Remove source files from registry.}
\usage{
removeRegistrySourceFiles(reg, src.files)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{src.files}{[\code{character}]\cr
Paths to remove from registry.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{src.files} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}},
  \code{\link{addRegistrySourceDirs}},
  \code{\link{addRegistrySourceFiles}},
  \code{\link{batchExport}}, \code{\link{batchUnexport}},
  \code{\link{loadExports}},
  \code{\link{removeRegistryPackages}},
  \code{\link{removeRegistrySourceDirs}},
  \code{\link{setRegistryPackages}}
}

