% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sanitizePath}
\alias{sanitizePath}
\title{Sanitize a path}
\usage{
sanitizePath(path, normalize = NULL)
}
\arguments{
\item{path}{[\code{character}]\cr
Vector of paths to sanitize.}

\item{normalize}{[\code{logical}]\cr
Do call \code{\link[base]{normalizePath}} on \code{path}.
Default is \code{NULL} which will only normalize if the path
is already an absolute path, i.e. relative paths will remain unchanged.}
}
\value{
\code{character} with sanitized paths.
}
\description{
Replaces backward slashes with forward slashes and optionally
normalizes the path.
}

