\name{sourceRegistryFiles}
\alias{sourceRegistryFiles}
\title{Source files in source directories specified in the registry.}
\usage{
  sourceRegistryFiles(reg, envir = .GlobalEnv)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{envir}{[\code{environment}]\cr Environment to
  source the files into. Default is the global
  environment.}
}
\value{
  Nothing.
}
\description{
  Source files in source directories specified in the
  registry.
}

