% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncRegistry.R
\name{syncRegistry}
\alias{syncRegistry}
\title{Syncronize staged queries into the registry.}
\usage{
syncRegistry(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
If the option \dQuote{staged.queries} is enabled, all communication from the nodes
to the master is done via files in the subdirectory \dQuote{pending} of the \code{file.dir}.
This function checks for such files and merges the information into the database.
Usually you do not have to call this function yourself.
}

