% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRegistry.R
\name{updateRegistry}
\alias{updateRegistry}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
updateRegistry(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}
}
\value{
[any]. Updated \code{\link{Registry}} or \code{FALSE} if no updates were performed.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

