% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Packages.R
\name{addRegistryPackages}
\alias{addRegistryPackages}
\title{Add packages to registry.}
\usage{
addRegistryPackages(reg, packages)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{packages}{[\code{character}]\cr
Packages to add to registry.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{packages} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistrySourceDirs}},
  \code{\link{addRegistrySourceFiles}},
  \code{\link{batchExport}}, \code{\link{batchUnexport}},
  \code{\link{loadExports}},
  \code{\link{removeRegistryPackages}},
  \code{\link{removeRegistrySourceDirs}},
  \code{\link{removeRegistrySourceFiles}},
  \code{\link{setRegistryPackages}}
}

