% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRegistryFiles.R
\name{addRegistrySourceFiles}
\alias{addRegistrySourceFiles}
\title{Add source files to registry.}
\usage{
addRegistrySourceFiles(reg, src.files, src.now = TRUE)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{src.files}{[\code{character}]\cr
Paths to add to registry.
See \code{\link{makeRegistry}}.}

\item{src.now}{[\code{logical(1)}]
Source files now on master?
Default is \code{TRUE}.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{src.files} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}},
  \code{\link{addRegistrySourceDirs}},
  \code{\link{batchExport}}, \code{\link{batchUnexport}},
  \code{\link{loadExports}},
  \code{\link{removeRegistryPackages}},
  \code{\link{removeRegistrySourceDirs}},
  \code{\link{removeRegistrySourceFiles}},
  \code{\link{setRegistryPackages}}
}

