\name{applyJobFunction}
\alias{applyJobFunction}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
applyJobFunction(reg, job)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{job}{[\code{\link{Job}}]\cr Job.}
}
\value{
[any]. Result of job.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

