% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyJobFunction.R
\name{applyJobFunction}
\alias{applyJobFunction}
\title{applyJobFunction
ONLY FOR INTERNAL USAGE.}
\usage{
applyJobFunction(reg, job, cache)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{job}{[\code{\link{Job}}]\cr
Job.}

\item{cache}{[\code{FileCache}]\cr
Instance of \code{\link[BBmisc]{makeFileCache}}.}
}
\value{
[any]. Result of job.
}
\description{
applyJobFunction
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

