% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{BatchJobs}
\alias{BatchJobs}
\alias{BatchJobs-package}
\title{The BatchJobs package}
\description{
The BatchJobs package
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/tudo-r/BatchJobs}}
  \item{Wiki:}{\url{https://github.com/tudo-r/BatchJobs/wiki}}
  \item{FAQ:}{\url{https://github.com/tudo-r/BatchJobs/wiki/FAQ}}
  \item{Configuration:}{\url{https://github.com/tudo-r/BatchJobs/wiki/Configuration}}
}
}

