% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{batchUnexport}
\alias{batchUnexport}
\title{Unload exported R objects.}
\usage{
batchUnexport(reg, what)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{what}{[\code{character}]\cr
Names of objects to remove.}
}
\value{
[\code{character}]. Invisibly returns a character vector of unexported objects.
}
\description{
Removes \code{RData} files from the \dQuote{exports} subdirectory of your \code{file.dir}
and thereby prevents loading on the slave.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}},
  \code{\link{addRegistrySourceDirs}},
  \code{\link{addRegistrySourceFiles}},
  \code{\link{batchExport}}, \code{\link{loadExports}},
  \code{\link{removeRegistryPackages}},
  \code{\link{removeRegistrySourceDirs}},
  \code{\link{removeRegistrySourceFiles}},
  \code{\link{setRegistryPackages}}
}

