% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{configuration}
\alias{.BatchJobs.R}
\alias{configuration}
\title{BatchJobs configuration.}
\description{
In order to understand how the package should be configured
please read
\url{https://github.com/tudo-r/BatchJobs/wiki/Configuration}.
}
\seealso{
Other conf: \code{\link{getConfig}},
  \code{\link{loadConfig}}, \code{\link{setConfig}}
}

