% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dbCreateJobDefTable}
\alias{dbCreateJobDefTable}
\title{ONLY FOR INTERNAL USAGE.}
\usage{
dbCreateJobDefTable(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}
}
\value{
Nothing.
}
\description{
ONLY FOR INTERNAL USAGE.
}
\keyword{internal}

