% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conf.R
\name{getConfig}
\alias{getConfig}
\title{Returns a list of BatchJobs configuration settings}
\usage{
getConfig()
}
\value{
\code{list} of current configuration variables with classs \dQuote{Config}.
}
\description{
Returns a list of BatchJobs configuration settings
}
\seealso{
Other conf: \code{\link{.BatchJobs.R}},
  \code{\link{configuration}}; \code{\link{loadConfig}};
  \code{\link{setConfig}}
}

