% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrorMessages.R
\name{getErrorMessages}
\alias{getErrorMessages}
\title{Get error messages of jobs.}
\usage{
getErrorMessages(reg, ids)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all jobs with errors.}
}
\value{
[\code{character}]. Error messages for jobs as character vector\cr
  \code{NA} if job has terminated successfully.
}
\description{
Get error messages of jobs.
}
\seealso{
Other debug: \code{\link{debugMulticore}},
  \code{\link{debugSSH}}, \code{\link{getJobInfo}},
  \code{\link{getLogFiles}}, \code{\link{grepLogs}},
  \code{\link{killJobs}}, \code{\link{resetJobs}},
  \code{\link{setJobFunction}}, \code{\link{showLog}},
  \code{\link{testJob}}
}

