% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getJobParamDf.R
\name{getJobParamDf}
\alias{getJobParamDf}
\title{Returns parameters for all jobs as the rows of a data.frame.}
\usage{
getJobParamDf(reg, ids)
}
\arguments{
\item{reg}{[\code{Registry}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all jobs.}
}
\value{
[\code{data.frame}]. Rows are named with job ids.
}
\description{
Returns parameters for all jobs as the rows of a data.frame.
}
\examples{
# see batchExpandGrid
}

