% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJob.R
\name{getJobs}
\alias{getJobs}
\title{Get jobs from registry by id.}
\usage{
getJobs(reg, ids, check.ids = TRUE)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all jobs.}

\item{check.ids}{[\code{logical(1)}]\cr
Check the job ids?
Default is \code{TRUE}.}
}
\value{
[list of \code{\link{Job}}].
}
\description{
Get jobs from registry by id.
}

