\name{getLogFiles}
\alias{getLogFiles}
\title{Get log file paths for jobs.}
\usage{
  getLogFiles(reg, ids)
}
\arguments{
  \item{reg}{[\code{\link{Registry}}]\cr Registry.}

  \item{ids}{[\code{integer}]\cr Ids of jobs.  Default is
  all jobs.}
}
\value{
  [\code{character}]. Vector of file paths to log files.
}
\description{
  Get log file paths for jobs.
}

