% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogFiles.R
\name{getLogFiles}
\alias{getLogFiles}
\title{Get log file paths for jobs.}
\usage{
getLogFiles(reg, ids)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{ids}{[\code{integer}]\cr
Ids of jobs.
Default is all jobs.}
}
\value{
[\code{character}]. Vector of file paths to log files.
}
\description{
Get log file paths for jobs.
}
\seealso{
Other debug: \code{\link{debugMulticore}},
  \code{\link{debugSSH}}, \code{\link{getErrorMessages}},
  \code{\link{getJobInfo}}, \code{\link{grepLogs}},
  \code{\link{killJobs}}, \code{\link{resetJobs}},
  \code{\link{setJobFunction}}, \code{\link{showLog}},
  \code{\link{testJob}}
}

