% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{loadConfig}
\alias{loadConfig}
\title{Load a specific configuration file.}
\usage{
loadConfig(conffile = ".BatchJobs.R")
}
\arguments{
\item{conffile}{[\code{character(1)}]\cr
Location of the configuration file to load.
Default is \dQuote{.BatchJobs.conf} in the current working directory.}
}
\value{
Invisibly returns a list of configuration settings.
}
\description{
Load a specific configuration file.
}
\seealso{
Other conf: \code{\link{configuration}},
  \code{\link{getConfig}}, \code{\link{setConfig}}
}

