% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadExports}
\alias{loadExports}
\title{Load exported R data objects.}
\usage{
loadExports(reg)
}
\arguments{
\item{reg}{[\code{Registry}]\cr
Registry.}
}
\value{
[\code{character}]. Invisibly returns a character vector of loaded objects.
}
\description{
Loads all \code{RData} files in the \dQuote{exports} subdirectory of your \code{file.dir}
and assigns the objects to the global environment.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}};
  \code{\link{addRegistrySourceDirs}};
  \code{\link{addRegistrySourceFiles}};
  \code{\link{batchExport}}; \code{\link{batchUnexport}};
  \code{\link{removeRegistryPackages}};
  \code{\link{removeRegistrySourceDirs}};
  \code{\link{removeRegistrySourceFiles}}
}

