% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResult.R
\name{loadResult}
\alias{loadResult}
\title{Loads a specific result file.}
\usage{
loadResult(reg, id, part = NA_character_, missing.ok = FALSE,
  impute.val = NULL)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry.}

\item{id}{[\code{integer(1)}]\cr
Id of job.}

\item{part}{[\code{character}]
Only useful for multiple result files, then defines which result file part(s) should be loaded.
\code{NA} means all parts are loaded, which is the default.}

\item{missing.ok}{[\code{logical(1)}]\cr
If \code{FALSE} an error is thrown if no result file is found.
Otherwise \code{NULL} is returned.
Default is \code{FALSE}.}

\item{impute.val}{[any]\cr
The value to return when no result is available.
Defaults to \code{NULL} (the previous behavior)}
}
\value{
[any]. Result of job.
}
\description{
Loads a specific result file.
}
\seealso{
\code{\link{reduceResults}}
}

