% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsLocal.R
\name{makeClusterFunctionsLocal}
\alias{makeClusterFunctionsLocal}
\title{Create cluster functions for sequential execution on local host.}
\usage{
makeClusterFunctionsLocal()
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
All jobs executed under these cluster functions are executed
sequentially, but in an independent, new R session.
That is, \code{submitJob} does not return until the
job has finished. The main use of this \code{ClusterFunctions}
implementation is to test and debug programs on a local computer.

Listing jobs returns an empty vector (as no jobs can be running when you call this)
and \code{killJob} returns at once (for the same reason).
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSLURM}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}}
}

