% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsMulticore.R
\name{makeClusterFunctionsMulticore}
\alias{makeClusterFunctionsMulticore}
\title{Use multiple cores on local Linux machine to spawn parallel jobs.}
\usage{
makeClusterFunctionsMulticore(ncpus = max(getOption("mc.cores",
  parallel::detectCores()) - 1L, 1L), max.jobs, max.load, nice,
  r.options = c("--no-save", "--no-restore", "--no-init-file",
  "--no-site-file"), script)
}
\arguments{
\item{ncpus}{[\code{integer(1)}]\cr
Number of VPUs of worker.
Default is to use all cores but one, where total number of cores
"available" is given by option \code{\link[base:options]{mc.cores}}
and if that is not set it is inferred by
\code{\link[parallel]{detectCores}}.}

\item{max.jobs}{[\code{integer(1)}]\cr
Maximal number of jobs that can run concurrently for the current registry.
Default is \code{ncpus}.}

\item{max.load}{[\code{numeric(1)}]\cr
Load average (of the last 5 min) at which the worker is considered occupied,
so that no job can be submitted.
Default is inferred by \code{\link[parallel]{detectCores}}, cf. argument \code{ncpus}.}

\item{nice}{[\code{integer(1)}]\cr
Process priority to run R with set via nice. Integers between -20 and 19 are allowed.
If missing, processes are not nice'd and the system default applies (usually 0).}

\item{r.options}{[\code{character}]
Options for R and Rscript, one option per element of the vector,
a la \dQuote{--vanilla}.
Default is \code{c("--no-save", "--no-restore", "--no-init-file", "--no-site-file")}.}

\item{script}{[\code{character(1)}]\cr
Path to helper bash script which interacts with the worker.
You really should not have to touch this, as this would imply that we have screwed up and
published an incompatible version for your system.
This option is only provided as a last resort for very experienced hackers.
Note that the path has to be absolute.
This is what is done in the package:
\url{https://github.com/tudo-r/BatchJobs/blob/master/inst/bin/linux-helper}
Default means to take it from package directory.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Jobs are spawned by starting multiple R sessions on the commandline
(similar like on true batch systems).
Packages \code{parallel} or \code{multicore} are not used in any way.
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsLocal}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSLURM}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}}
}

