% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeRegistryPackages}
\alias{removeRegistryPackages}
\title{Remove packages from registry.}
\usage{
removeRegistryPackages(reg, packages)
}
\arguments{
\item{reg}{[\code{Registry}]\cr
Registry.}

\item{packages}{[\code{character}]\cr
Packages to remove from registry.}
}
\value{
[\code{\link{Registry}}]. Changed registry.
}
\description{
Mutator function for \code{packages} in \code{\link{makeRegistry}}.
}
\seealso{
Other exports: \code{\link{addRegistryPackages}};
  \code{\link{addRegistrySourceDirs}};
  \code{\link{addRegistrySourceFiles}};
  \code{\link{batchExport}}; \code{\link{batchUnexport}};
  \code{\link{loadExports}};
  \code{\link{removeRegistrySourceDirs}};
  \code{\link{removeRegistrySourceFiles}}
}

