% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizePath.R
\name{sanitizePath}
\alias{sanitizePath}
\title{Sanitize a path}
\usage{
sanitizePath(path, make.absolute = TRUE, normalize.absolute = FALSE)
}
\arguments{
\item{path}{[\code{character}]\cr
Vector of paths to sanitize.}

\item{make.absolute}{[\code{logical}]\cr
If \code{TRUE} convert to an absolute path.}

\item{normalize.absolute}{[\code{logical}]\cr
Also call \code{\link[base]{normalizePath}} on absolute paths?
This will immediately resolve symlinks.}
}
\value{
\code{character} with sanitized paths.
}
\description{
Replaces backward slashes with forward slashes and optionally
normalizes the path.
}

